Fix TeamStatistics now module, StatHelper super class
