+++ lib/stat_tracker.rb
@@ -1,7 +1,10 @@
 require 'csv'
 require 'pry'
+require_relative 'stat_helper'
+require_relative 'team_statistics'
 
-class StatTracker
+class StatTracker < StatHelper
+  include TeamStatistics
   attr_reader :games, :teams, :game_teams
   def initialize(games, teams, game_teams)
     @games = games
@@ -88,7 +91,7 @@ class StatTracker
   def average_goals_by_season
     season_goal_averages = Hash.new
     total_goals_per_season.each do |season, goals|
-      season_goal_averages[season] = (goals / count_of_games_by_season[season]).round(2) 
+      season_goal_averages[season] = (goals / count_of_games_by_season[season]).round(2)
     end
     season_goal_averages
   end
@@ -299,15 +302,6 @@ class StatTracker
       team[:team_id] == least_accurate_team_id
     end
     least_accurate_team[:teamname]
-  end
-
-  # Helper method is used in tackles_by_team
-  def games_by_season
-    @games_by_season_hash = Hash.new([])
-    @games.each do |game|
-      @games_by_season_hash[game[:season]] += [game[:game_id]]
-    end
-    @games_by_season_hash
   end
 
   # Helper method is used in most_tackles & fewest_tackles
