+++ lib/team_statistics.rb
@@ -1,19 +1,7 @@
 require 'csv'
 require 'pry'
 
-class TeamStatistics < StatTracker
-  attr_reader :games, :teams, :game_teams
-  def initialize(games, teams, game_teams)
-    super(games, teams, game_teams)
-  end
-
-  def self.from_csv(csv_hash)
-    games_input = CSV.read(csv_hash[:games], headers: true, header_converters: :symbol)
-    teams_input = CSV.read(csv_hash[:teams], headers: true, header_converters: :symbol)
-    game_teams_input = CSV.read(csv_hash[:game_teams], headers: true, header_converters: :symbol)
-    stats_tracker = TeamStatistics.new(games_input, teams_input, game_teams_input)
-  end
-
+module TeamStatistics
   # Original method from Iteration 2
   def team_info(team_id)
     team_info_hash = Hash.new
