+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 require "simplecov"
 SimpleCov.start
 
+require_relative 'stat_helper_spec'
 require_relative 'stat_tracker_spec'
 require_relative 'team_statistics_spec'
